requirements.txt
pdfplumber
PyPDF2
pdfminer.six
pytesseract
pillow
pdf2image
opencv-python
pymupdf
